Created a basic shopping list item component. The Shopping List now dynamically creates cards based on the shoppingList state prop
