--- src/components/ShoppingListContainer.js
@@ -1,13 +1,0 @@
-import { useState } from "react"
-
-
-const ShoppingListContainer = (props) => {
-    return (
-        <section>
-
-        </section>
-    )
-
-}
-
-export default ShoppingListContainer
